perf: getLatestProjects method returns projects with a type order
